+++ lib/llm_cost_tracker/retention.rb
@@ -21,17 +21,6 @@ module LlmCostTracker
         deleted
       end
 
-      def prune_invoice_imports(older_than:, now: Time.now.utc)
-        cutoff = resolve_cutoff(older_than, now)
-        require_relative "ledger"
-        return 0 unless LlmCostTracker::ProviderInvoiceImport.table_exists?
-
-        LlmCostTracker::ProviderInvoiceImport
-          .where(state: %w[completed failed])
-          .where(finished_at: ...cutoff)
-          .delete_all
-      end
-
       def prune_inbox(older_than:, now: Time.now.utc)
         cutoff = resolve_cutoff(older_than, now)
         require_relative "ingestion"